Some tests for util.find_amount()
